Implement protocol specific scheduler queries
